- Added interfaces for service classes to implement. This will allow for creation of completely new services in the future, without breaking the rest of the code.
